考勤模块-加班-调班-请假

+++ store-system-common/src/main/java/com/store/system/model/attendance/WorkOverTime.java
@@ -37,7 +37,7 @@ public class WorkOverTime implements Serializable {
     /**
      * 申请时间
      */
-    private double applyTime;
+    private long applyTime;
     /**
      * 加班时长(小时)
      */
